move routes from subfolder of controllers to main folder
