--- src/controllers/routes/productRoute/getStockUse.js
@@ -1,18 +1,0 @@
-const { selectStockConsumeByProductId } = require('../../../models/products');
-
-const getStockUse = async (req, res, next) => {
-  const productId = req.params.productId;
-  
-  try {
-    const itemList = await selectStockConsumeByProductId({productId});
-    const response = {
-      flag: 'success',
-      message: 'get ingredient',
-      data: itemList,
-    }
-    res.status(200).json(response);
-  } catch (e) {
-    next(e);
-  }
-}
-module.exports = getStockUse;
