--- src/routes/auth.js
@@ -1,16 +1,0 @@
-const router = require('express').Router();
-
-const lemondb = require('../config/lemondb');
-
-// controllers
-const login = require('../controllers/authControllers/login');
-const register = require('../controllers/authControllers/register');
-const check = require('../controllers/authControllers/check');
-
-router.post('/login', login(lemondb));
-// demo plan
-router.post('/register', register(lemondb));
-
-router.get('/check', check);
-
-module.exports = router;
