+++ src/routes/authRoute/index.js
