+++ src/routes/getConPath.js
